+++ src/index.ts
@@ -1,13 +1,16 @@
 export {
     Botist,
+    IAdapter,
     IResponse as IRequestResponse,
     ISuccess as IRequestSuccess,
     IError as IRequestError,
+    IPoll,
 } from './Botist';
 export {
     IMessage,
     ITextMessage,
     IImageMessage,
+    IPollMessage,
     MessageType,
 } from './Message.t';
 export { MainScene, IScene } from './MainScene';
@@ -18,6 +21,6 @@ export { Telegram } from './Telegram/Adapter';
 export { Response as IResponse } from './Response';
 export { MessageMiddleware, ISubscriber, ISubscriberCallback } from './Middlewares/Message';
 export { CatchMiddleware, ICatchMiddleware } from './Middlewares/Catch';
-export { SubscriberContext } from './SubscriberContext';
+export { SceneContext } from './SceneContext';
 export { IEvent } from './Events/Event';
 export * from './Events/SceneEvents';